[REFACTOR] 번잡한 헤더 병합 및 Makefile 주석 추가
