+++ Makefile
@@ -6,23 +6,30 @@
 #    By: seongmik <seongmik@student.42seoul.kr>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/01/03 19:51:24 by seongmik          #+#    #+#              #
-#    Updated: 2024/01/07 03:18:35 by seongmik         ###   ########.fr        #
+#    Updated: 2024/01/08 14:19:06 by seongmik         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
+
+#--------------------------------default flags---------------------------------#
 NAME							=	minishell
 CC								=	cc
 CFLAGS							=	-Wall -Wextra -Werror
 
-REDLINE_FALG					=	-lreadline
 
-#-----------------------this flags is for M1 mac-----------------------#
-#before make, you should install readline with brew using below command#
-#brew install readline                                                 #
+#--------------------------------readline flag---------------------------------#
+REDLINE_FLAG					=	-lreadline
+
+
+#---------------------------this flags is for M1 mac---------------------------#
+#----before make, you should install readline with brew using below command----#
+#----brew install readline-----------------------------------------------------#
 M1_MAC_READLINE_FLAGS			=	-L/opt/homebrew/opt/readline/lib
 M1_MAC_READLINE_OBJ_FLAGS		=	-I/opt/homebrew/opt/readline/include
-#----------------------------------------------------------------------#
+#------------------------------------------------------------------------------#
 
+
+#----------------------------------sources-------------------------------------#
 INCLUDES 						:=	-I./includes
 SRCS_DIR 						:=	srcs
 LIB_DIR							:=	srcs/libft
@@ -37,23 +44,28 @@ SRCS							:=	$(SRCS_DIR)/minishell.c \
 									$(SRCS_DIR)/builtin_exit.c \
 									$(SRCS_DIR)/ft_env.c \
 									$(SRCS_DIR)/ft_env_util.c \
-									$(SRCS_DIR)/executor_command.c \
+									$(SRCS_DIR)/execute_command.c \
 									$(SRCS_DIR)/string_util.c \
 									$(SRCS_DIR)/command.c \
 									$(SRCS_DIR)/heredoc.c \
 									$(SRCS_DIR)/setting_signal.c \
 									$(SRCS_DIR)/shell_error.c \
-									$(SRCS_DIR)/word_expand.c
+									$(SRCS_DIR)/word_expand.c \
+									$(SRCS_DIR)/word_expand_util.c
 
+
+#----------------------------------objects-------------------------------------#
 OBJS							=	$(SRCS:.c=.o)
 
+
+#-----------------------------------rules--------------------------------------#
 all : $(NAME)
 
 $(NAME) : $(OBJS)
 	@echo "Compiling..."
 	@$(MAKE) -C $(LIB_DIR);
 	@cp $(LIB_DIR)/libft.a libft.a
-	@$(CC) $(CFLAGS) -o $@ $^ libft.a $(REDLINE_FALG) $(M1_MAC_READLINE_FLAGS) #-fsanitize=address
+	@$(CC) $(CFLAGS) -o $@ $^ libft.a $(REDLINE_FLAG) $(M1_MAC_READLINE_FLAGS) #-fsanitize=address
 	@echo "Done !"
 
 %.o : %.c
@@ -76,4 +88,6 @@ fclean : clean
 
 re : fclean all
 
+
+#-----------------------------------phony--------------------------------------#
 .PHONY: all clean fclean re
