--- includes/exit_status.h
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   exit_status.h                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: seongmik <seongmik@student.42seoul.kr>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/03 16:12:55 by seongmik          #+#    #+#             */
-/*   Updated: 2024/01/04 15:04:19 by seongmik         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef EXIT_STATUS_H
-# define EXIT_STATUS_H
-
-# define EXIT_SUCCESS	0
-# define EXIT_FAILURE	1
-# define SUCCESS		0
-# define FAILURE		1
-
-#endif
