+++ srcs/execute_command.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   executor_command.c                                 :+:      :+:    :+:   */
+/*   execute_command.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: seongmik <seongmik@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/04 14:44:57 by seongmik          #+#    #+#             */
-/*   Updated: 2024/01/07 15:02:17 by seongmik         ###   ########.fr       */
+/*   Updated: 2024/01/08 14:18:53 by seongmik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
